Added:checking maximum in float with test case

+++ src/main/java/com/generics/GetMaximum.java
@@ -2,14 +2,16 @@ package com.generics;
 
 public class GetMaximum {
 
-    public static Integer getmaximum(Integer a,Integer b,Integer c)
+    public static Float getmaximum(Float a,Float b,Float c)
     {
-        Integer max=a;
+        Float max=a;
         if(b.compareTo(max) > 0 && b.compareTo(c) > 0)
             max=b;
         else if(c.compareTo(max) > 0)
             max=c;
 
+
+        //System.out.println(max);
         return max;
     }
